Pixy code
After pondering last night, I decided to give it another try. This
should get 64 bytes of data from the pixy, parse it until it finds the
beginning of the frame (two 0xaa55 shorts in a row), save the next 6
bytes along with the block number to an array, saves that array to an
array, then finds the next block, does the same, then returns the array
of two arrays so we can then use that data. For simplicity I removed
some parts, like the ability to recognize color-code blocks, the
servo-moving function, and a couple others. I hope it works.